Move symwalk into fsutil
